Conflicts:
	ffxivminion/ffxiv_common_cne.lua
	ffxivminion/ffxiv_helpers.lua
	ffxivminion/ffxiv_task_duty.lua
	minionlib/languages.lua

+++ minionlib/ml_task.lua
@@ -12,6 +12,9 @@ ml_task.auxiliary = false
 ml_task.process_elements = {}
 ml_task.overwatch_elements = {}
 ml_task.breakUpdate = false
+ml_task.delayTime = 0
+ml_task.delayTimer = 0
+ml_task.preserveSubtasks = false
 
 -- These functions are NOT overwritten in derived tasks
 
@@ -24,6 +27,11 @@ end
 
 function ml_task:isValid()
 	return self.valid
+end
+
+function ml_task:Invalidate()
+	ml_debug(self.name.."->Invalidate(Fail)()")
+	self.valid = false
 end
 
 function ml_task:hasCompleted()
@@ -70,9 +78,11 @@ function ml_task:Update()
             return TS_FAILED
         end
         if ( self:hasCompleted() ) then
-			ml_debug(self.name.." has succeeded")
-            self:DeleteSubTasks()
-            return TS_SUCCEEDED
+			if(TimeSince(self.delayTime) > self.delayTimer) then
+				ml_debug(self.name.." has succeeded")
+				self:DeleteSubTasks()
+				return TS_SUCCEEDED
+			end
         end
         local taskRet = nil
 		
@@ -82,13 +92,19 @@ function ml_task:Update()
 		end
 		
 		if(self:ProcessOverWatch()) then
-			ml_debug(self.name.."->ProcessOverWatch executed an effect, breaking loop")
-			--process overwatch element requested to break update loop
-			--only delete subtask if we didn't just add it via our overwatch cne
-			if (self.subtask ~= nil and (currentSubtaskName ~= nil or self.subtask.name == currentSubtaskName)) then
-				self:DeleteSubTasks()
+			if (not self.preserveSubtasks) then
+				ml_debug(self.name.."->ProcessOverWatch executed an effect, breaking loop")
+				--process overwatch element requested to break update loop
+				--only delete subtask if we didn't just add it via our overwatch cne
+
+				if (self.subtask ~= nil and (currentSubtaskName ~= nil or self.subtask.name == currentSubtaskName)) then
+					self:DeleteSubTasks()
+				end
+				
+				break
+			else
+				self.preserveSubtasks = false
 			end
-			break
 		end
 		
         if ( self.subtask ~= nil ) then
@@ -104,7 +120,12 @@ function ml_task:Update()
         else
 			ml_debug(self.name.."->Process()")
 			gFFXIVMinionTask = self.name
-            continueUpdate = self:Process()
+			if(TimeSince(self.delayTime) > self.delayTimer) then
+				continueUpdate = self:Process()
+			else
+				ml_debug("Delaying Process for "..tostring(self.delayTimer - (ml_global_information.Now - self.delayTime)).."ms")
+				continueUpdate = false
+			end
         end
     end
 	ml_debug(self.name.."->Update() returning")
@@ -119,9 +140,7 @@ function ml_task:Process()
     if (TableSize(self.process_elements) > 0) then
 		ml_cne_hub.clear_queue()
 		ml_cne_hub.eval_elements(self.process_elements)
-		if (self:superClass() and TableSize(self:superClass().process_elements) > 0) then
-			ml_cne_hub.eval_elements(self:superClass().process_elements)
-		end
+
 		ml_cne_hub.queue_to_execute()
 		ml_cne_hub.execute()
 		return false
@@ -137,12 +156,24 @@ function ml_task:ProcessOverWatch()
 		ml_debug(self.name.."->ProcessOverWatch()")
 		ml_cne_hub.clear_queue()
 		ml_cne_hub.eval_elements(self.overwatch_elements)
-		if (self:superClass() and TableSize(self:superClass().overwatch_elements) > 0) then
-			ml_cne_hub.eval_elements(self:superClass().overwatch_elements)
-		end
+
 		ml_cne_hub.queue_to_execute()
 		return ml_cne_hub.execute()
 	end
+end
+
+function ml_task:SetDelay(delayTimer)
+	--ignore 0 case
+	if(	type(delayTimer) == "number" and delayTimer > 0) then
+		self.delayTime = ml_global_information.Now
+		self.delayTimer = delayTimer
+	elseif(	type(delayTimer) == "number" and delayTimer < 0) then
+		ml_error("Invalid delaytimer input - negative number")
+	end
+end
+
+function ml_task:IsDelayed()
+	return TimeSince(self.delayTime) < self.delayTimer
 end
 
 --These functions ARE overwritten in derived tasks
@@ -194,7 +225,7 @@ function ml_task:task_complete_eval()
 end
 
 function ml_task:task_complete_execute()
-
+    self.completed = true
 end
 
 function ml_task:task_fail_eval()
@@ -202,7 +233,7 @@ function ml_task:task_fail_eval()
 end
 
 function ml_task:task_fail_execute()
- 
+    self.completed = true
 end
 
 function ml_task:AddTaskCheckCEs()
